refactor: merge yuanbao_markdown.py + yuanbao_api.py into yuanbao.py
Consolidate markdown chunking utilities and sign-token API logic into
the main yuanbao.py module, reducing yuanbao file count from 6 to 4.
Pure code relocation with no logic changes. All 39 tests pass.
